requirements.txt
faiss-cpu>=1.7.4
transformers>=4.36.2
langchain>=0.3.18
openai>=1.6.1
numpy>=1.24.3
pydantic>=2.7.0
python-dotenv==1.0.0
langchain-openai>=0.0.10
langchain-community>=0.3.17
torch>=1.13.0
requests==2.31.0